+++ src/sine.c
@@ -261,7 +261,7 @@ void two_stage_pitch_refinement(MODEL *model, COMP Sw[])
   if (model->Wo > TWO_PI/P_MIN)
     model->Wo = TWO_PI/P_MIN;
 
-  model->L = floor(PI/model->Wo);
+  model->L = floorf(PI/model->Wo);
 }
 
 /*---------------------------------------------------------------------------*\
@@ -369,7 +369,7 @@ void estimate_amplitudes(MODEL *model, COMP Sw[], COMP W[], int est_phase)
         /* Estimate phase of harmonic, this is expensive in CPU for
            embedded devicesso we make it an option */
 
-        model->phi[m] = atan2(Sw[b].imag,Sw[b].real);
+        model->phi[m] = atan2f(Sw[b].imag,Sw[b].real);
     }
   }
 }
@@ -423,8 +423,8 @@ float est_voicing_mbe(
 	Am.real = 0.0;
 	Am.imag = 0.0;
 	den = 0.0;
-	al = ceil((l - 0.5)*Wo*FFT_ENC/TWO_PI);
-	bl = ceil((l + 0.5)*Wo*FFT_ENC/TWO_PI);
+	al = ceilf((l - 0.5)*Wo*FFT_ENC/TWO_PI);
+	bl = ceilf((l + 0.5)*Wo*FFT_ENC/TWO_PI);
 
 	/* Estimate amplitude of harmonic assuming harmonic is totally voiced */
 
@@ -616,10 +616,10 @@ void synthesise(
     */
     for(l=1; l<=model->L; l++) {
 	for(i=0,j=-N+1; i<N-1; i++,j++) {
-	    Sw_[FFT_DEC-N+1+i].real += 2.0*model->A[l]*cos(j*model->Wo*l + model->phi[l]);
+	    Sw_[FFT_DEC-N+1+i].real += 2.0*model->A[l]*cosf(j*model->Wo*l + model->phi[l]);
 	}
  	for(i=N-1,j=0; i<2*N; i++,j++)
-	    Sw_[j].real += 2.0*model->A[l]*cos(j*model->Wo*l + model->phi[l]);
+	    Sw_[j].real += 2.0*model->A[l]*cosf(j*model->Wo*l + model->phi[l]);
     }	
 #endif
